Add standalone kitchen sink
